Agent_Evaluation/requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.0
attrs==25.3.0
azure-ai-evaluation==1.3.0
azure-ai-projects==1.0.0b7
azure-core==1.32.0
azure-identity==1.21.0
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-storage-blob==12.25.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.11
filelock==3.18.0
filetype==1.2.0
fixedint==0.1.6
Flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.69.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
openai==1.66.5
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
promptflow-core==1.17.2
promptflow-devkit==1.17.2
promptflow-tracing==1.17.2
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydash==7.0.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==310
pywin32-ctypes==0.2.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
